requirements.txt
streamlit
pandas
openpyxl
duckdb
ollama
sqlglot
PyYAML
huggingface-hub
llama-cpp-python